% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsea.R
\name{plotSEA}
\alias{plotSEA}
\title{topSEA}
\usage{
plotSEA(object, by = "TDP.estimate", threshold = 0.005, n = 20)
}
\arguments{
\item{object}{A SEA-chart object which is the output of \code{SEA} function}

\item{by}{the Variable which will we mapped.
It should be either the TDP estimate or TDP bound.The default is TDP bound.}

\item{threshold}{A real number between 0 and 1. Which will be used as
a visual aid to distinguish significant pathways}

\item{n}{Integer. Number of rows from SEA-chart object to be plotted.}
}
\value{
Returns a plot of SEA_chart according to the selected arguments
}
\description{
returns a plotof SEA-chart which illustrates
proportion of discoveries per pathway.
}
\examples{
#See the examples for \code{\link{SEA}}

}
\references{
Mitra Ebrahimpoor, Pietro Spitali, Kristina Hettne, Roula Tsonaka, Jelle Goeman,
Simultaneous Enrichment Analysis of all Possible Gene-sets: Unifying Self-Contained
and Competitive Methods, Briefings in Bioinformatics,bbz074
}
\seealso{
\code{\link{SEA}}
}
\author{
Mitra Ebrahimpoor

\email{m.ebrahimpoor@lumc.nl}
}
